+++ packages/layer-list/esbuild/global.d.ts
@@ -1,3 +1,13 @@
 // These constants here are added by the base esbuild config
 
 declare const PLUGIN_VERSION: string;
+
+declare module '*.svg?react' {
+  import * as React from 'react';
+
+  const ReactComponent: React.FunctionComponent<
+    React.SVGProps<SVGSVGElement> & { title?: string }
+  >;
+
+  export default ReactComponent;
+}
